Extract the source map server to its own file
